services/llm-tutor/requirements.txt
# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# AI/ML Libraries
langchain==0.0.336
langchain-community==0.0.6
openai==1.3.5
huggingface-hub==0.17.3
transformers==4.35.2
sentence-transformers==2.2.2
torch==2.1.1
accelerate==0.24.1
unstructured[md]==0.10.30

# Vector Database & Search
pymilvus==2.3.4
elasticsearch==8.11.0
redis==5.0.1
hiredis==2.2.3

# Audio Processing
openai-whisper==20231117
TTS==0.20.6
soundfile==0.12.1
librosa==0.10.1

# Database & ORM
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0

# Security & Privacy
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
click==8.1.7

# Production
gunicorn==21.2.0
structlog==23.2.0
python-dotenv==1.0.0
